requirements.txt
requests
pandas
pymongo
python-dotenv
streamlit
plotly
streamlit-autorefresh
